+++ scripts/build_gitlab/kw.sh
@@ -36,11 +36,7 @@ INSTALL_DIR=`pwd`/install
 mkdir ospray_build
 cd ospray_build
 
-export rkcommon_DIR=$INSTALL_DIR
-export glfw3_DIR=$INSTALL_DIR
-export embree_DIR=$INSTALL_DIR
-export openvkl_DIR=$INSTALL_DIR
-
+export CMAKE_PREFIX_PATH=${INSTALL_DIR}
 
 cmake -DISPC_EXECUTABLE=$INSTALL_DIR/bin/ispc -DTBB_ROOT=$INSTALL_DIR ../..